pyproject.toml
[tool.poetry]
name = "TLE"
version = "0.2.0"
description = ""
authors = ["meooow25 <soumiksarkar.3120@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"
matplotlib="3.8.4"
seaborn = "^0.10.1"
lxml = "^4.6"
pillow = "^9.0"
pycairo = "^1.19.1"
PyGObject = "^3.34.0"
aiocache = "^0.11.1"
jishaku = "^2.5.2"
ujson = "^5.9.0"
google-generativeai = "^0.3.2"
google-api-python-client = "^2.134.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
markovify = "^0.9.4"
spacy = "^3.7.5"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
